feat(#138): rename generation package -> bytecode package
