+++ src/main/java/org/eolang/jeo/representation/bytecode/BytecodeClass.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation.generation;
+package org.eolang.jeo.representation.bytecode;
 
 import com.jcabi.xml.XML;
 import java.io.PrintWriter;
