+++ src/main/java/org/eolang/jeo/representation/bytecode/BytecodeInstruction.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation.generation;
+package org.eolang.jeo.representation.bytecode;
 
 import java.util.Arrays;
 import java.util.List;
